jpegxl/interop-2026/features/progressive-dc-observable.html

<!DOCTYPE html>
<title>JPEG XL interop: progressive codestream decode</title>
<link rel="help" href="https://github.com/web-platform-tests/interop-jpegxl">
<meta name="assert" content="Decodes progressive JPEG XL fixture and verifies expected dimensions.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/interop-helpers.js"></script>
<script>

promise_test(async () => {
  const img = await loadImage('../resources/progressive_ac.jxl');
  assert_equals(img.naturalWidth, 2268);
  assert_equals(img.naturalHeight, 1512);
}, 'Progressive codestream fixture decodes and renders final frame.');

</script>
